requirements.txt
tensorflow==1.14.0
keras==2.3.1
opencv-python-headless
flask
numpy
pandas
pillow
gunicorn